# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank API
  description: Bankcard API made as a test task
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
  /api/transactions/withdraw:
    post:
      tags:
        - Transactions
      summary: Withdraw money to user's card
      description: Withdraws money from user's card with currency conversion
      operationId: withdraw
      requestBody:
        description: Parameters needed for one card operation
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OneCardOperationDto"
        required: true
      responses:
        "200":
          description: Card with updated balance returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "401":
          description: User is not authorized for the action
        "403":
          description: Card is not active or user is not an owner of the card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/transactions/transfer:
    post:
      tags:
        - Transactions
      summary: Transfer money from one user's card to another
      description: Transfers money from card to card with currency conversion
      operationId: transfer
      requestBody:
        description: Transfer parameters
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequestDto"
        required: true
      responses:
        "200":
          description: Card with updated balance returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "401":
          description: User is not authorized for the action
        "403":
          description: One of the cards is not active or user is not an owner of the
            card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: One of the cards not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/transactions/deposit:
    post:
      tags:
        - Transactions
      summary: Deposit money to user's card
      description: Deposits money to user's card with currency conversion
      operationId: deposit
      requestBody:
        description: Parameters needed for one card operation
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OneCardOperationDto"
        required: true
      responses:
        "200":
          description: Card with updated balance returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "401":
          description: User is not authorized for the action
        "403":
          description: Card is not active or user is not an owner of the card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/users/{user-id}:
    get:
      tags:
        - Admin - Users
      summary: Get a user
      description: Returns the user with given id
      operationId: getUserById
      parameters:
        - name: user-id
          in: path
          description: UUID of the user
          required: true
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        "200":
          description: User was returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "401":
          description: User is not authorized for the action
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Admin - Users
      summary: Delete a user
      description: Deletes user with given id
      operationId: deleteUserById
      parameters:
        - name: user-id
          in: path
          description: UUID of the user
          required: true
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        "204":
          description: User was deleted
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Admin - Users
      summary: Make user an admin
      description: Changes the role of a given user to admin
      operationId: makeAdmin
      parameters:
        - name: user-id
          in: path
          description: UUID of the user
          required: true
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        "204":
          description: User role was changed to admin
        "401":
          description: User is not authorized for the action
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/admin/users:
    get:
      tags:
        - Admin - Users
      summary: Get the list of users
      description: Returns all existing users
      operationId: getUsers
      responses:
        "200":
          description: List of UserDtos was returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    OneCardOperationDto:
      required:
        - currency
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        cardNumber:
          maxLength: 19
          minLength: 16
          type: string
        currency:
          type: string
          enum:
            - RUB
            - USD
            - EUR
        amount:
          type: number
    CardDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - ACTIVE
            - LOCKED
            - EXPIRED
            - PENDING_LOCK
        currency:
          type: string
          enum:
            - RUB
            - USD
            - EUR
        balance:
          type: number
        masked_number:
          type: string
        owner_id:
          type: string
          format: uuid
        expiration_date:
          type: string
          format: date
    ContentDisposition:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        filename:
          type: string
        charset:
          type: string
        size:
          type: integer
          format: int64
          deprecated: true
        creationDate:
          type: string
          format: date-time
          deprecated: true
        modificationDate:
          type: string
          format: date-time
          deprecated: true
        readDate:
          type: string
          format: date-time
          deprecated: true
        attachment:
          type: boolean
        formData:
          type: boolean
        inline:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        body:
          $ref: "#/components/schemas/ProblemDetail"
        headers:
          type: object
          properties:
            connection:
              type: array
              items:
                type: string
            contentDisposition:
              $ref: "#/components/schemas/ContentDisposition"
            acceptCharset:
              type: array
              items:
                type: string
            location:
              type: string
              format: uri
            empty:
              type: boolean
            host:
              type: object
              properties:
                hostString:
                  type: string
                address:
                  type: object
                  properties:
                    hostAddress:
                      type: string
                    address:
                      type: array
                      items:
                        type: string
                        format: byte
                    hostName:
                      type: string
                    linkLocalAddress:
                      type: boolean
                    multicastAddress:
                      type: boolean
                    anyLocalAddress:
                      type: boolean
                    loopbackAddress:
                      type: boolean
                    siteLocalAddress:
                      type: boolean
                    mcglobal:
                      type: boolean
                    mcnodeLocal:
                      type: boolean
                    mclinkLocal:
                      type: boolean
                    mcsiteLocal:
                      type: boolean
                    mcorgLocal:
                      type: boolean
                    canonicalHostName:
                      type: string
                port:
                  type: integer
                  format: int32
                unresolved:
                  type: boolean
                hostName:
                  type: string
            all:
              type: object
              additionalProperties:
                type: string
              writeOnly: true
            lastModified:
              type: integer
              format: int64
            date:
              type: integer
              format: int64
            contentLength:
              type: integer
              format: int64
            contentType:
              $ref: "#/components/schemas/MediaType"
            ifModifiedSince:
              type: integer
              format: int64
            origin:
              type: string
            range:
              type: array
              items:
                $ref: "#/components/schemas/HttpRange"
            allow:
              uniqueItems: true
              type: array
              items:
                $ref: "#/components/schemas/HttpMethod"
            cacheControl:
              type: string
            contentLanguage:
              type: object
              properties:
                language:
                  type: string
                displayName:
                  type: string
                country:
                  type: string
                variant:
                  type: string
                script:
                  type: string
                unicodeLocaleAttributes:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                unicodeLocaleKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                displayLanguage:
                  type: string
                displayScript:
                  type: string
                displayCountry:
                  type: string
                displayVariant:
                  type: string
                extensionKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                iso3Language:
                  type: string
                iso3Country:
                  type: string
            etag:
              type: string
            accessControlAllowOrigin:
              type: string
            accessControlAllowMethods:
              type: array
              items:
                $ref: "#/components/schemas/HttpMethod"
            accessControlAllowHeaders:
              type: array
              items:
                type: string
            accessControlExposeHeaders:
              type: array
              items:
                type: string
            accessControlAllowCredentials:
              type: boolean
            accessControlMaxAge:
              type: integer
              format: int64
            accessControlRequestMethod:
              $ref: "#/components/schemas/HttpMethod"
            accessControlRequestHeaders:
              type: array
              items:
                type: string
            accept:
              type: array
              items:
                $ref: "#/components/schemas/MediaType"
            acceptLanguageAsLocales:
              type: array
              items:
                type: object
                properties:
                  language:
                    type: string
                  displayName:
                    type: string
                  country:
                    type: string
                  variant:
                    type: string
                  script:
                    type: string
                  unicodeLocaleAttributes:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  unicodeLocaleKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  displayLanguage:
                    type: string
                  displayScript:
                    type: string
                  displayCountry:
                    type: string
                  displayVariant:
                    type: string
                  extensionKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  iso3Language:
                    type: string
                  iso3Country:
                    type: string
            acceptPatch:
              type: array
              items:
                $ref: "#/components/schemas/MediaType"
            bearerAuth:
              type: string
              writeOnly: true
            expires:
              type: integer
              format: int64
            ifMatch:
              type: array
              items:
                type: string
            ifNoneMatch:
              type: array
              items:
                type: string
            ifUnmodifiedSince:
              type: integer
              format: int64
            pragma:
              type: string
            upgrade:
              type: string
            vary:
              type: array
              items:
                type: string
            acceptLanguage:
              type: array
              items:
                type: object
                properties:
                  range:
                    type: string
                  weight:
                    type: number
                    format: double
            basicAuth:
              type: string
              writeOnly: true
          additionalProperties:
            type: array
            items:
              type: string
        statusCode:
          $ref: "#/components/schemas/HttpStatusCode"
        typeMessageCode:
          type: string
        detailMessageCode:
          type: string
        detailMessageArguments:
          type: array
          items:
            type: object
        titleMessageCode:
          type: string
    HttpMethod:
      type: object
    HttpRange:
      type: object
    HttpStatusCode:
      type: object
      properties:
        is4xxClientError:
          type: boolean
        error:
          type: boolean
        is5xxServerError:
          type: boolean
        is1xxInformational:
          type: boolean
        is2xxSuccessful:
          type: boolean
        is3xxRedirection:
          type: boolean
    MediaType:
      type: object
      properties:
        type:
          type: string
        subtype:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        qualityValue:
          type: number
          format: double
        wildcardType:
          type: boolean
        wildcardSubtype:
          type: boolean
        subtypeSuffix:
          type: string
        charset:
          type: string
        concrete:
          type: boolean
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    TransferRequestDto:
      type: object
      properties:
        cardFromId:
          type: string
          format: uuid
        cardFromNumber:
          maxLength: 19
          minLength: 16
          type: string
        cardToId:
          type: string
          format: uuid
        cardToNumber:
          maxLength: 19
          minLength: 16
          type: string
        amount:
          type: number
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT